Make toolbars be seperate classes
